remove print statements where it should be logging, unify some exceptions
