+++ pyrit/score/human_in_the_loop_scorer.py
@@ -185,11 +185,10 @@ class HumanInTheLoopScorer(Scorer):
             import tkinter as tk
             from tkinter import simpledialog
         except ImportError as e:
-            print(
+            raise RuntimeError(
                 "To use HumanInTheLoopScorer, you need to install tkinter. "
                 "See https://stackoverflow.com/a/74607246 for more information."
-            )
-            raise e
+            ) from e
         root = tk.Tk()
         root.withdraw()
         user_input = simpledialog.askstring("Score Prompt", message).strip()
